chore(code/starknet): Change signature scheme from ECDSA to Ed25519
